Update overlap.py
Remove references to upstream and downstream as that can be confusing due to the fact that genes can be in sense or antisense orientation. Left and right is most accurate for now. We will assign sense and antisense values, swapping the density calculations post processing. 

Add TODO for validating window on the left rho calculation.